Address reviewer request
Conjoin so that a single test "suite" exists for CSI tests.
Excise filesystem default opts probing code, it is unreliable.
