+++ csi/src/mount.rs
@@ -1,9 +1,6 @@
 //! Utility functions for working with mountpoints
 
-use std::process::Command;
-
 use proc_mounts::MountIter;
-use run_script::ScriptOptions;
 use sys_mount::{unmount, FilesystemType, Mount, MountFlags, UnmountFlags};
 
 // Information about a mounted filesystem.
@@ -114,152 +111,18 @@ pub fn mount_opts_compare(m1: &[String], m2: &[String], ro: bool) -> bool {
 pub fn probe_filesystems() -> Vec<Fs> {
     let mut filesystems = Vec::new();
     // the first filesystem is the default one
-    let supported_fs = ["xfs", "ext4"];
 
-    for fsname in supported_fs.iter() {
-        match probe_filesystem(fsname) {
-            Ok(opts) => filesystems.push(Fs {
-                name: (*fsname).to_string(),
-                defaults: opts,
-            }),
-            Err(err) => {
-                warn!("Filesystem {} will not be available: {}", fsname, err)
-            }
-        }
-    }
+    filesystems.push(Fs {
+        name: "xfs".to_string(),
+        defaults: vec![],
+    });
+
+    filesystems.push(Fs {
+        name: "ext4".to_string(),
+        defaults: vec![],
+    });
 
     filesystems
-}
-
-fn probe_filesystem(fsname: &str) -> Result<Vec<String>, String> {
-    let mi = MountInfo {
-        source: format!("/tmp/fs-{}.img", fsname),
-        dest: format!("/tmp/fs_default_{}", fsname),
-        opts: Vec::new(),
-    };
-    prepare_mount(&mi)?;
-
-    let opts = probe_defaults(fsname, &mi)?;
-    cleanup_mount(&mi).unwrap_or_else(|e| {
-        warn!("{}", e);
-    });
-    Ok(opts)
-}
-
-fn prepare_mount(mi: &MountInfo) -> Result<(), String> {
-    // im using run_script as its a very nice macro to not "exec" something
-    // rather it actually builds a proper script.
-    let mut options = ScriptOptions::new();
-    options.exit_on_error = true;
-
-    // truncate in busybox does not support units, so we express the size in
-    // bytes
-    let (code, _stdout, stderr) = run_script!(
-        r#"
-        truncate -s 67108864 $1
-        mkdir -p $2
-        "#,
-        vec!(mi.source.clone(), mi.dest.clone()),
-        &options
-    )
-    .unwrap();
-
-    if code == 0 {
-        return Ok(());
-    }
-    Err(format!(
-        "Failed to set up default mount options files: {}",
-        stderr
-    ))
-}
-
-fn cleanup_mount(mi: &MountInfo) -> Result<(), String> {
-    let mut options = ScriptOptions::new();
-    options.exit_on_error = true;
-
-    // stdout of clean up script is always empty
-    let (code, _stdout, stderr) = run_script!(
-        r#"
-        rmdir $2
-        rm -f $1
-        "#,
-        &vec!(mi.source.clone(), mi.dest.clone()),
-        &options
-    )
-    .unwrap();
-
-    if code == 0 {
-        return Ok(());
-    }
-    Err(format!(
-        "Failed to cleanup default mount options files: {}",
-        stderr
-    ))
-}
-
-// After some research, it turns out that default mount options
-// are depending on various CONFIG_XXXX options during kernel
-// config. Depending on FS (sigh) the defaults can be determined
-// by an util like tune2fs. However, this requires you to have a
-// filesystem of that type to begin with... (how useful)
-//
-// I have found no way to determine these options by digging through
-// sysfs so here is a hack. I feel bad about this. I hate to do this
-// but I've given up. Linux won, there you have it.
-fn probe_defaults(fsname: &str, mi: &MountInfo) -> Result<Vec<String>, String> {
-    let output = Command::new(format!("mkfs.{}", fsname))
-        .arg(&mi.source)
-        .output()
-        .expect("Failed to execute mkfs command");
-    if !output.status.success() {
-        return Err(format!(
-            "Failed to mkfs {} fs: {}",
-            fsname,
-            String::from_utf8(output.stderr).unwrap()
-        ));
-    }
-    trace!(
-        "Output of mkfs.{} command: {}",
-        fsname,
-        String::from_utf8(output.stdout).unwrap()
-    );
-
-    let output = Command::new("mount")
-        .arg(&mi.source)
-        .arg(&mi.dest)
-        .output()
-        .expect("Failed to execute mount command");
-    if !output.status.success() {
-        return Err(format!(
-            "Failed to mount {} fs: {}",
-            fsname,
-            String::from_utf8(output.stderr).unwrap()
-        ));
-    }
-
-    let mut options = match_mount(None, Some(&mi.dest), true).unwrap().opts;
-
-    let output = Command::new("umount")
-        .arg("-vf")
-        .arg(&mi.dest)
-        .output()
-        .expect("Failed to execute umount command");
-
-    if !output.status.success() {
-        error!(
-            "Failed to unmount {} probe fs: {}",
-            fsname,
-            String::from_utf8(output.stderr).unwrap()
-        );
-    }
-
-    options
-        .iter()
-        .position(|n| n == "rw")
-        .map(|e| options.remove(e));
-
-    info!("Default {} mount options: {}", fsname, options.join(","));
-    Ok(options)
 }
 
 /// Mount filesystem
