+++ test.sh
@@ -6,8 +6,6 @@ cargo build --all
 ( cd mayastor && cargo test -- --test-threads=1 )
 ( cd mayastor-test && npm install && ./node_modules/mocha/bin/mocha test_cli.js )
 ( cd mayastor-test && ./node_modules/mocha/bin/mocha test_replica.js )
-( cd mayastor-test && ./node_modules/mocha/bin/mocha test_csi_iscsi.js )
-( cd mayastor-test && ./node_modules/mocha/bin/mocha test_csi_nbd.js )
 ( cd mayastor-test && ./node_modules/mocha/bin/mocha test_csi.js )
 ( cd mayastor-test && ./node_modules/mocha/bin/mocha test_nexus.js )
 ( cd mayastor-test && ./node_modules/mocha/bin/mocha test_rebuild.js )
